requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.7
psycopg2-binary>=2.9.5
pydantic>=1.10.7
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
alembic>=1.10.2
pandas>=2.0.0
openpyxl>=3.1.2
httpx>=0.24.0
jinja2>=3.1.2
pdfkit>=1.0.0
pytest>=7.3.1
crewai>=0.1.0
pydantic>=2.5.1 
email-validator>=1.1.2